requirements.txt
ping3
requests
bs4
fake-useragent
colorama
python-dateutil